apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.28.0",
    "fastapi>=0.104.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "google-cloud-pubsub>=2.18.0",
    "google-genai>=1.29.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.0",
    "sqlmodel>=0.0.8",
    "supabase>=2.0.0",
    "types-requests>=2.32.4.20250611",
    "uvicorn[standard]>=0.23.0",
    "cryptography>=41.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pytest>=9.0.2",
]
